refactor(corebundle): 消费 cellmodules/replaydeps，删除已迁移的 Redis 复用胶水代码 [#825][#2017]
- redis.go: 删除已迁移到 replaydeps 的 env const、类型、vars、函数
  （loadRedisConfigFromEnv / parseClusterAddrs / buildRedisClient /
  buildServiceNonceStore / buildConsumerClaimer）；保留 HTTP 幂等 store 相关代码
- shared_deps_build.go: 删除 sharedReplayDeps struct 与 buildSharedReplayDeps func
- shared_deps.go: 将 buildSharedReplayDeps(ctx, topo, clk) 替换为
  replaydeps.Resolve(ctx, clk, topo)，添加 cellmodules/replaydeps import；
  下游字段访问 replay.RedisClient / .ConsumerClaimer / .NonceStore 不变
- redis_test.go: 删除已迁至 cellmodules/replaydeps/replaydeps_test.go 的测试；
  将 bundle_test.go 仍引用的 fakeDistributedClaimer 迁出至新文件
  redis_testhelper_test.go
- controlplane.go / bundle_options.go / auth_integration_test.go: 更新注释中的
  函数名引用为 replaydeps.Resolve
